Move schema tests to col_v dir
