+++ tests/integration/acp/dac/link_schema/accept_multi_resources_on_dri_test.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/sourcenetwork/defradb/tests/action"
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
-	schemaUtils "github.com/sourcenetwork/defradb/tests/integration/schema"
+	schemaUtils "github.com/sourcenetwork/defradb/tests/integration/collection_version"
 )
 
 func TestACP_LinkSchema_WithMultipleResources_AcceptSchema(t *testing.T) {
