+++ tests/integration/acp/dac/link_schema/accept_same_resource_on_diff_schemas_test.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/sourcenetwork/defradb/tests/action"
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
-	schemaUtils "github.com/sourcenetwork/defradb/tests/integration/schema"
+	schemaUtils "github.com/sourcenetwork/defradb/tests/integration/collection_version"
 )
 
 func TestACP_LinkSchema_UseSameResourceOnDifferentSchemas_AcceptSchemas(t *testing.T) {
