+++ tests/integration/collection_version/client_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package schema
+package collection_version
 
 import (
 	"testing"
