+++ tests/integration/collection_version/nil_type_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package schema
+package collection_version
 
 import (
 	"testing"
@@ -17,40 +17,16 @@ import (
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
 )
 
-func TestEncryptFieldsForCreateMutation(t *testing.T) {
+func TestSchema_WithMissingType_Errors(t *testing.T) {
 	test := testUtils.TestCase{
-
 		Actions: []any{
 			&action.AddSchema{
 				Schema: `
 					type User {
-						name: String
-						age:  Int
+						name:
 					}
 				`,
-			},
-			testUtils.IntrospectionRequest{
-				Request: `
-				{
-				  __type(name: "UserField") {
-				    name
-				    kind
-				    enumValues {
-				      name
-				    }
-				  }
-				}
-				`,
-				ContainsData: map[string]any{
-					"__type": map[string]any{
-						"kind": "ENUM",
-						"name": "UserField",
-						"enumValues": []any{
-							map[string]any{"name": "name"},
-							map[string]any{"name": "age"},
-						},
-					},
-				},
+				ExpectedError: "field type not specified. Object: User, Field: name",
 			},
 		},
 	}
