+++ tests/integration/collection_version/updates/copy/field/with_introspection_test.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/sourcenetwork/defradb/tests/action"
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
-	introspectionUtils "github.com/sourcenetwork/defradb/tests/integration/schema"
+	introspectionUtils "github.com/sourcenetwork/defradb/tests/integration/collection_version"
 )
 
 func TestSchemaUpdatesCopyFieldIntrospectionWithRemoveIDAndReplaceName(t *testing.T) {
